Switch to oxlint (#3024)
* Convert from eslint to oxlint

* Use req'd args in resource transactional fns

* Remove TS signature overloads in Resource module

These are purely TS type metadata, but the linter doesn't like them. TS can infer types on these that seem better, and if we do want to type them differently than that, I think we should use union types (or similar) instead of these overloads. This should be a backwards-compatible change as the actual signature of the functions does not change.

* Fix all the other issues oxlint identified

* Enable es2024 env in oxlint

...and update globals accordingly

* Reformat oxlint config file

* Fix JSDoc params in transactionBroadcast/addSubscription

* Disable the same rules we had disabled in eslint

* Only specify oxlint rules that differ from defaults

* Remove eslint

* Remove globals package that is not needed by oxlint